kata: update shimv2 command to support runtime-rs
The rust implementation for the kata shim is behaving the same as the go one,
except on some parameter validation.
Typically, the "publish events"-related parameters need to be not empty.

This is an issue with cri-o, which doesn't support this API and doesn't provide
the parameters - making them empty.

This commit ensure that the parameters are either set to something acceptable,
or removed from the environment altogether.

Signed-off-by: Julien Ropé <jrope@redhat.com>